chore: cleanup pass — ruff fixes, dedupe helper, stale-comment hunt
A safe-bounded full-repo cleanup.

Ruff (19 auto + unsafe fixes applied across the tree)
- RUF046 unnecessary int() around round() (round already returns int).
- RUF005 collection-literal concatenation → unpacked literal
  (chans + (x,) → (*chans, x)).
- B905 zip without explicit strict= → strict=False on the call sites
  that intentionally tolerate length mismatch.
- RUF007 zip(a[:-1], a[1:]) → itertools.pairwise.
- B007 unused loop var renamed to _-prefixed.
- RUF022 __all__ alphabetised.
- I001 import blocks sorted.
- UP037 unnecessary string-quoted return annotations stripped.

Dead / stale
- shortcuts_dialog: removed Shift+C "Toggle contact sheet" listing
  (the shortcut was retired with the contact-sheet feature in v1.2;
  users hitting it would have nothing happen). Added the missing
  Compare section (W / Ctrl+W / right-click drag).
- transport.py: dropped "+contact-sheet" leftover comments around
  channel_selection_changed signal and channel button construction.
- app._init_channel_state docstring: dropped contact-sheet history
  paragraph (kept the relevant per-layer fallback note).

Factorisation
- New _qbool(raw, default=False) helper in preferences.py — folds
  Windows REG_DWORD ints + .conf "true"/"false" strings into a
  Python bool. Replaces 5 inline isinstance(raw, str)+lower() blocks
  in Preferences (display_timecode, channel_labels_visible, …) and
  the duplicated _coerce_bool that lived in save_frame_handler. Test
  imports updated to point at the new home.

Validation
- 864 tests pass (5 deselected pre-existing flaky), ruff clean across
  src/.

No behavioural changes — pure cleanup. The bigger app.py extraction
(annotation handlers → annotation_handler.py, ~400 lines) is left
for a follow-up since it needs more dedicated test coverage to land
safely.

+++ src/img_player/compare/compose.py
@@ -86,7 +86,7 @@ def compose(
         # interactive seam-drag, where every mouse-move triggers
         # one of these per frame.
         h, w = a.shape[:2]
-        split = int(round(w * max(0.0, min(1.0, float(seam)))))
+        split = round(w * max(0.0, min(1.0, float(seam))))
         out = np.empty_like(a)
         if split > 0:
             out[:, :split] = a[:, :split]
@@ -97,7 +97,7 @@ def compose(
         return out
     if mode == MODE_HORIZONTAL:
         h, w = a.shape[:2]
-        split = int(round(h * max(0.0, min(1.0, float(seam)))))
+        split = round(h * max(0.0, min(1.0, float(seam))))
         out = np.empty_like(a)
         if split > 0:
             out[:split] = a[:split]
